pyproject.toml
[project]
name = "ENGENHARIA-DE-DADOS"
version = "0.1.0"
description = ""
authors = [
    {name = "ALAN AMORIM",email = "alanamorim28@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "streamlit (>=1.55.0,<2.0.0)",
    "plotly-express (>=0.4.1,<0.5.0)"
]

[tool.poetry]
packages = [{include = "poetry init "}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)"
]
